+++ packages/core/src/scxml.ts
@@ -389,12 +389,6 @@ function toConfig(
         target: target ? `#${target}` : undefined
       };
     }
-    case 'final': {
-      return {
-        ...nodeJson.attributes,
-        type: 'final'
-      };
-    }
     default:
       break;
   }
@@ -448,13 +442,6 @@ function toConfig(
         );
 
         return events.map((event) => {
-          if (event === 'done.invoke') {
-            throw new Error(
-              'done.invoke gets often used in SCXML tests as inexact event descriptor.' +
-                " As long as this stay unimplemented or done.invoke doesn't get a specialcased while converting throw when seeing it to avoid tests using this to pass by accident."
-            );
-          }
-
           const targets = getAttribute(value, 'target');
           const internal = getAttribute(value, 'type') === 'internal';
 
@@ -543,6 +530,7 @@ function toConfig(
           }
         : undefined),
       ...(parallel ? { type: 'parallel' } : undefined),
+      ...(nodeJson.name === 'final' ? { type: 'final' } : undefined),
       ...(stateElements.length
         ? {
             states: mapValues(states, (state, key) =>
@@ -557,7 +545,7 @@ function toConfig(
     };
   }
 
-  return { id };
+  return { id, ...(nodeJson.name === 'final' ? { type: 'final' } : undefined) };
 }
 
 export interface ScxmlToMachineOptions {
@@ -604,7 +592,8 @@ function scxmlToMachine(
   return Machine({
     ...toConfig(machineElement, '(machine)', options),
     context,
-    delimiter: options.delimiter
+    delimiter: options.delimiter,
+    scxml: true
   });
 }