+++ packages/core/src/stateUtils.ts
@@ -241,15 +241,73 @@ export const isStateId = (str: string) => str[0] === STATE_IDENTIFIER;
 
 export function getCandidates<TEvent extends EventObject>(
   stateNode: StateNode<any, TEvent>,
-  eventName: TEvent['type'] | NullEvent['type'] | '*'
+  receivedEventType: TEvent['type'] | NullEvent['type'],
+  /**
+   * If `true`, will use SCXML event partial token matching semantics
+   * without the need for the ".*" suffix
+   */
+  partialMatch: boolean = false
 ): Array<TransitionDefinition<any, TEvent>> {
-  const transient = eventName === NULL_EVENT;
+  const transient = receivedEventType === NULL_EVENT;
   const candidates = stateNode.transitions.filter((transition) => {
-    const sameEventType = transition.eventType === eventName;
-    // null events should only match against eventless transitions
-    return transient
-      ? sameEventType
-      : sameEventType || transition.eventType === WILDCARD;
+    const { eventType } = transition;
+    // First, check the trivial case: event names are exactly equal
+    if (eventType === receivedEventType) {
+      return true;
+    }
+
+    // Transient transitions can't match non-transient events
+    if (transient) {
+      return false;
+    }
+
+    // Then, check if transition is a wildcard transition,
+    // which matches any non-transient events
+    if (eventType === WILDCARD) {
+      return true;
+    }
+
+    if (!partialMatch && !eventType.endsWith('.*')) {
+      return false;
+    }
+
+    if (!IS_PRODUCTION) {
+      warn(
+        !/.*\*.+/.test(eventType),
+        `Wildcards can only be the last token of an event descriptor (e.g., "event.*") or the entire event descriptor ("*"). Check the "${eventType}" event.`
+      );
+    }
+
+    const partialEventTokens = eventType.split('.');
+    const eventTokens = receivedEventType.split('.');
+
+    for (
+      let tokenIndex = 0;
+      tokenIndex < partialEventTokens.length;
+      tokenIndex++
+    ) {
+      const partialEventToken = partialEventTokens[tokenIndex];
+      const eventToken = eventTokens[tokenIndex];
+
+      if (partialEventToken === '*') {
+        const isLastToken = tokenIndex === partialEventTokens.length - 1;
+
+        if (!IS_PRODUCTION) {
+          warn(
+            isLastToken,
+            `Infix wildcards in transition events are not allowed. Check the "${eventType}" event.`
+          );
+        }
+
+        return isLastToken;
+      }
+
+      if (partialEventToken !== eventToken) {
+        return false;
+      }
+    }
+
+    return true;
   });
 
   return candidates;
@@ -678,10 +736,9 @@ export function getStateNodes<TContext, TEvent extends EventObject>(
   }
 
   const childStateKeys = keys(stateValue);
-  const childStateNodes: Array<StateNode<
-    TContext,
-    TEvent
-  >> = childStateKeys
+  const childStateNodes: Array<
+    StateNode<TContext, TEvent>
+  > = childStateKeys
     .map((subStateKey) => getStateNode(stateNode, subStateKey))
     .filter(Boolean);
 
@@ -1336,10 +1393,9 @@ export function microstep<TContext, TEvent extends EventObject>(
 function selectEventlessTransitions<TContext, TEvent extends EventObject>(
   state: State<TContext, TEvent>
 ): Transitions<TContext, TEvent> {
-  const enabledTransitions: Set<TransitionDefinition<
-    TContext,
-    TEvent
-  >> = new Set();
+  const enabledTransitions: Set<
+    TransitionDefinition<TContext, TEvent>
+  > = new Set();
 
   const atomicStates = state.configuration.filter(isAtomicStateNode);