+++ packages/xstate-vue/test/UseMachine.vue
@@ -36,7 +36,7 @@ const fetchMachine = Machine<typeof context, any>({
           actions: assign({
             data: (_, e) => e.data
           }),
-          cond: (_, e) => e.data.length
+          guard: (_, e) => e.data.length
         }
       }
     },