+++ packages/xstate-inspect/CHANGELOG.md
@@ -1,5 +1,74 @@
 # @xstate/inspect
 
+## 0.4.1
+
+### Patch Changes
+
+- [`d9282107`](https://github.com/davidkpiano/xstate/commit/d9282107b931b867d9cd297ede71b55fe11eb74d) [#1800](https://github.com/davidkpiano/xstate/pull/1800) Thanks [@davidkpiano](https://github.com/davidkpiano)! - Fixed a bug where services were not being registered by the inspect client, affecting the ability to send events to inspected services.
+
+## 0.4.0
+
+### Minor Changes
+
+- [`63ba888e`](https://github.com/davidkpiano/xstate/commit/63ba888e19bd2b72f9aad2c9cd36cde297e0ffe5) [#1770](https://github.com/davidkpiano/xstate/pull/1770) Thanks [@davidkpiano](https://github.com/davidkpiano)! - It is now easier for developers to create their own XState inspectors, and even inspect services offline.
+
+  A **receiver** is an actor that receives inspector events from a source, such as `"service.register"`, `"service.state"`, `"service.event"`, etc. This update includes two receivers:
+
+  - `createWindowReceiver` - listens to inspector events from a parent window (for both popup and iframe scenarios)
+  - 🚧 `createWebSocketReceiver` (experimental) - listens to inspector events from a WebSocket server
+
+  Here's how it works:
+
+  **Application (browser) code**
+
+  ```js
+  import { inspect } from '@xstate/inspect';
+
+  inspect(/* options */);
+
+  // ...
+
+  interpret(someMachine, { devTools: true }).start();
+  ```
+
+  **Inspector code**
+
+  ```js
+  import { createWindowReceiver } from '@xstate/inspect';
+
+  const windowReceiver = createWindowReceiver(/* options? */);
+
+  windowReceiver.subscribe((event) => {
+    // here, you will receive events like:
+    // { type: "service.register", machine: ..., state: ..., sessionId: ... }
+    console.log(event);
+  });
+  ```
+
+  The events you will receive are `ParsedReceiverEvent` types:
+
+  ```ts
+  export type ParsedReceiverEvent =
+    | {
+        type: 'service.register';
+        machine: StateMachine<any, any, any>;
+        state: State<any, any>;
+        id: string;
+        sessionId: string;
+        parent?: string;
+        source?: string;
+      }
+    | { type: 'service.stop'; sessionId: string }
+    | {
+        type: 'service.state';
+        state: State<any, any>;
+        sessionId: string;
+      }
+    | { type: 'service.event'; event: SCXML.Event<any>; sessionId: string };
+  ```
+
+  Given these events, you can visualize the service machines and their states and events however you'd like.
+
 ## 0.3.0
 
 ### Minor Changes